+++ packages/react-native-gesture-handler/src/handlers/gestures/installUIRuntimeBindings.web.ts
@@ -1,0 +1,8 @@
+// UI runtime bindings are a native-only concept — there is no separate UI
+// worklet runtime on web, and `getUIRuntimeHolder` from react-native-worklets
+// throws when called there.
+export function installUIRuntimeBindings(
+  _getUIRuntimeHolder: (() => object) | undefined
+) {
+  // no-op
+}
